ghost_in_the_c/requirements.txt
google-genai
python-dotenv
pydantic